django-rest-swagger/requirements.txt
# REST Api
Django==1.10.6
djangorestframework==3.8.2
coreapi==2.3.0
openapi-codec==1.3.1
simplejson==3.9.0

# Multi Env Testing
tox==2.4.1
coverage==4.2
# Auto comp.
ipdb==0.10.1

# Docs
mkdocs==0.15.3

# PostGreSQL
psycopg2==2.7.3.2
# Multiple instances
gunicorn==19.7.0
# Static files
whitenoise==3.3.0


# Image
pillow==5.2.0